+++ dacite/CHANGELOG.md
@@ -21,6 +21,7 @@
    - `core::SubpassDescription`
    - `core::RenderPassCreateInfo`
    - `core::RenderPassBeginInfo`
+   - `khr_swapchain::SwapchainCreateInfoKhr`
 
  - The builder `core::SpecializationInfoBuilder` has been removed. Entries to a
    `core::SpecializationInfo` can now be added conveniently with `core::SpecializationInfo::push`.